requirements.txt
Flask==2.3.2
scikit-learn==1.3.0
nltk==3.8.1
pyspellchecker==0.7.1
